# system-design-challenges-50/day05/docs/api_spec.yaml
openapi: 3.0.0
info:
  title: Scalable Requirements Tracker API
  description: API for managing system requirements with version control
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.requirements-tracker.example.com/api/v1
    description: Production server

paths:
  /auth/login:
    post:
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials

  /auth/register:
    post:
      summary: Register new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Successful registration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Invalid input

  /requirements/:
    get:
      summary: List requirements
      operationId: listRequirements
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of requirements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RequirementResponse'
    post:
      summary: Create requirement
      operationId: createRequirement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementCreate'
      responses:
        '200':
          description: Created requirement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '400':
          description: Invalid input

  /requirements/{requirement_id}:
    get:
      summary: Get requirement by ID
      operationId: getRequirement
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Requirement details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '404':
          description: Requirement not found
    put:
      summary: Update requirement
      operationId: updateRequirement
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementUpdate'
      responses:
        '200':
          description: Updated requirement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementResponse'
        '404':
          description: Requirement not found
    delete:
      summary: Delete requirement
      operationId: deleteRequirement
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Requirement deleted
        '404':
          description: Requirement not found

  /versions/:
    post:
      summary: Create version
      operationId: createVersion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionCreate'
      responses:
        '200':
          description: Created version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionResponse'
        '400':
          description: Invalid input

  /versions/{version_id}:
    get:
      summary: Get version by ID
      operationId: getVersion
      parameters:
        - name: version_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Version details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionResponse'
        '404':
          description: Version not found
    put:
      summary: Update version
      operationId: updateVersion
      parameters:
        - name: version_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VersionUpdate'
      responses:
        '200':
          description: Updated version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionResponse'
        '404':
          description: Version not found
    delete:
      summary: Delete version
      operationId: deleteVersion
      parameters:
        - name: version_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Version deleted
        '404':
          description: Version not found

  /versions/requirement/{requirement_id}:
    get:
      summary: List versions for requirement
      operationId: listVersions
      parameters:
        - name: requirement_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VersionResponse'

  /projects/:
    get:
      summary: List projects
      operationId: listProjects
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProjectResponse'
    post:
      summary: Create project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '200':
          description: Created project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          description: Invalid input

  /projects/{project_id}:
    get:
      summary: Get project by ID
      operationId: getProject
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found
    put:
      summary: Update project
      operationId: updateProject
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          description: Project not found
    delete:
      summary: Delete project
      operationId: deleteProject
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Project deleted
        '404':
          description: Project not found

components:
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string

    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_active:
          type: boolean

    RequirementBase:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        status:
          type: string
          enum: [draft, in_progress, completed]
          default: draft
        project_id:
          type: integer

    RequirementCreate:
      allOf:
        - $ref: '#/components/schemas/RequirementBase'

    RequirementUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [draft, in_progress, completed]

    RequirementResponse:
      allOf:
        - $ref: '#/components/schemas/RequirementBase'
        - type: object
          properties:
            id:
              type: integer
            version:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time

    VersionBase:
      type: object
      properties:
        requirement_id:
          type: integer
        version_number:
          type: integer
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        status:
          type: string
          enum: [draft, in_progress, completed]
          default: draft

    VersionCreate:
      allOf:
        - $ref: '#/components/schemas/VersionBase'

    VersionUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [draft, in_progress, completed]

    VersionResponse:
      allOf:
        - $ref: '#/components/schemas/VersionBase'
        - type: object
          properties:
            id:
              type: integer
            created_at:
              type: string
              format: date-time

    ProjectBase:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    ProjectCreate:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string

    ProjectResponse:
      allOf:
        - $ref: '#/components/schemas/ProjectBase'
        - type: object
          properties:
            id:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []